Change int to size_t in randomString

+++ libs/generator/generator.cpp
@@ -30,7 +30,7 @@ namespace generator {
     return (uint8_t)random_number(32, 126 + 1);
   }
 
-  std::string randomString(int len) {
+  std::string randomString(size_t len) {
     std::string str(len, 0);
     std::generate_n(
         str.begin(), len, []() { return 'a' + std::rand() % ('z' - 'a' + 1); });
